Move /bundles folder to /framework/bundle
